requirements.txt
amqp==5.3.1
asgiref==3.8.1
astroid==3.3.8
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
celery==5.4.0
cffi==1.17.1
channels==4.2.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==44.0.1
daphne==4.1.2
dill==0.3.9
Django==5.0.12
django-bootstrap5==24.3
et_xmlfile==2.0.0
hyperlink==21.0.0
hypothesis==6.127.1
idna==3.10
incremental==24.7.2
iniconfig==2.0.0
isort==6.0.0
kombu==5.4.2
mccabe==0.7.0
mssql-django==1.5
numpy==2.2.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.50
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pylint==3.3.4
pyodbc==5.2.0
pyOpenSSL==25.0.0
pytest==8.3.4
pytest-django==4.10.0
python-dateutil==2.9.0.post0
pytz==2025.1
redis==5.2.1
service-identity==24.2.0
six==1.17.0
sortedcontainers==2.4.0
sqlparse==0.5.3
tomlkit==0.13.2
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.1
vine==5.1.0
wcwidth==0.2.13
zope.interface==7.2
gunicorn
channels-redis